+++ lab/SeedingWeedingOutAndDestroyingStartup/NSeed/Filtering/FullNameContainsSeedableFilter.cs
@@ -1,0 +1,26 @@
+using NSeed.MetaInfo;
+using System.Linq;
+
+namespace NSeed.Filtering
+{
+    /// <summary>
+    /// TODO.
+    /// </summary>
+    public class FullNameContainsSeedableFilter : ISeedableFilter
+    {
+        private readonly string[] partsOfSeedableFullName;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="FullNameContainsSeedableFilter"/> class.
+        /// TODO.
+        /// </summary>
+        /// <param name="partsOfSeedableFullName">TODO.</param>
+        public FullNameContainsSeedableFilter(params string[] partsOfSeedableFullName)
+        {
+            this.partsOfSeedableFullName = partsOfSeedableFullName;
+        }
+
+        /// <inheritdoc/>
+        public bool Accepts(SeedableInfo seedableInfo) => partsOfSeedableFullName.Any(part => seedableInfo.FullName.IndexOf(part, System.StringComparison.InvariantCultureIgnoreCase) >= 0);
+    }
+}