+++ lab/SeedingWeedingOutAndDestroyingStartup/NSeed/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the "NSeed" NuGet package will be documented in this file
 
 The format of the file is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [0.3.0-lab2] - Unreleased 
+### Added
+All added features are in the experimental phase and lab-quality:
+- `ISeedableFilter` interface with concrete implementations: `FullNameEqualsSeedableFilter`, `FullNameContainsSeedableFilter`, `AcceptAllSeedableFilter`.
+
 ## [0.3.0-lab1] - 2020-04-11
 ### Added
 All added features are in the experimental phase and lab-quality: